# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: PR/PO Approval System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchase-orders:
    get:
      operationId: getPurchaseOrders
      tags: [purchase-orders]
      summary: Get purchase orders
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, issued, receiving, received]
      responses:
        "200":
          description: List of purchase orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrderListResponse"

    post:
      operationId: createPurchaseOrder
      tags: [purchase-orders]
      summary: Create purchase order from PR
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePurchaseOrderRequest"
      responses:
        "201":
          description: PO created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchase-orders/{id}:
    get:
      operationId: getPurchaseOrderById
      tags: [purchase-orders]
      summary: Get PO by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: PO found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"

    put:
      operationId: updatePurchaseOrder
      tags: [purchase-orders]
      summary: Update PO
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePurchaseOrderRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchase-orders/{id}/issue:
    post:
      operationId: issuePurchaseOrder
      tags: [purchase-orders]
      summary: Issue a PO
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
  /purchase-orders/{id}/receive:
    post:
      operationId: receivePurchaseOrder
      tags: [purchase-orders]
      summary: Receive a PO
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Received
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseOrder"
...
components:
  schemas:
...
    PoItem:
      type: object
      properties:
        id:
          type: integer
        poId:
          type: integer
        prItemId:
          type: integer
          nullable: true
        name:
          type: string
        qty:
          type: number
        unit:
          type: string
        negotiatedPrice:
          type: number
        totalPrice:
          type: number
      required:
        - id
        - poId
        - name
        - qty
        - unit
        - negotiatedPrice
        - totalPrice
    PoItemInput:
      type: object
      properties:
        prItemId:
          type: integer
          nullable: true
        name:
          type: string
        qty:
          type: number
        unit:
          type: string
        negotiatedPrice:
          type: number
      required:
        - name
        - qty
        - unit
        - negotiatedPrice
    PurchaseOrder:
      type: object
      properties:
        id:
          type: integer
        poNumber:
          type: string
        prId:
          type: integer
        prNumber:
          type: string
        supplier:
          type: string
        status:
          type: string
          enum: [draft, issued, receiving, received]
        totalAmount:
          type: number
        notes:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/PoItem"
        createdById:
          type: integer
        createdByName:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - poNumber
        - prId
        - prNumber
        - supplier
        - status
        - totalAmount
        - items
        - createdById
        - createdByName
        - createdAt
        - updatedAt
    PurchaseOrderListResponse:
      type: object
      properties:
        purchaseOrders:
          type: array
          items:
            $ref: "#/components/schemas/PurchaseOrder"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - purchaseOrders
        - total
        - page
        - limit
    CreatePurchaseOrderRequest:
      type: object
      properties:
        prId:
          type: integer
        supplier:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PoItemInput"
        notes:
          type: string
          nullable: true
      required:
        - prId
        - supplier
        - items
    UpdatePurchaseOrderRequest:
      type: object
      properties:
        supplier:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/PoItemInput"
        notes:
          type: string
          nullable: true
...